Create FunSpecSerializer/Deserializer to many methods in JsonXExtensionsSpecBuilder, still needs more refactoring, clean up some static check issues - still a lot left
