+++ core/src/main/java/com/infinum/jsonapix/core/JsonApiModel.kt
@@ -4,8 +4,8 @@ import com.infinum.jsonapix.core.resources.DefaultLinks
 import com.infinum.jsonapix.core.resources.Links
 import com.infinum.jsonapix.core.resources.Meta
 
-@Suppress("UNCHECKED_CAST")
-abstract class JsonApiModel {
+@Suppress("UNCHECKED_CAST", "TooManyFunctions")
+open class JsonApiModel {
     private var rootLinks: Links? = null
     private var resourceLinks: Links? = null
     private var relationshipLinks: Map<String, Links?>? = null
@@ -29,16 +29,16 @@ abstract class JsonApiModel {
 
     fun rootLinks(): DefaultLinks? = rootLinks as? DefaultLinks
 
-    fun <T: Links> rootLinks(): T? = rootLinks as? T
+    fun <T : Links> rootLinks(): T? = rootLinks as? T
 
     fun resourceLinks(): DefaultLinks? = resourceLinks as? DefaultLinks
 
-    fun <T: Links> resourceLinks(): T? = resourceLinks as? T
+    fun <T : Links> resourceLinks(): T? = resourceLinks as? T
 
     fun relationshipsLinks(): Map<String, DefaultLinks?>? = relationshipLinks as? Map<String, DefaultLinks?>
 
     @JvmName("customRelationshipLinks")
-    fun <T: Links> relationshipsLinks(): Map<String, T?>? = relationshipLinks as? Map<String, T?>
+    fun <T : Links> relationshipsLinks(): Map<String, T?>? = relationshipLinks as? Map<String, T?>
 
-    fun <T: Meta> meta() = meta as? T
+    fun <T : Meta> meta() = meta as? T
 }
