+++ core/src/main/java/com/infinum/jsonapix/core/resources/DefaultLinks.kt
