+++ core/src/main/java/com/infinum/jsonapix/core/resources/Meta.kt
